use module sweetalert2 on delete and add simply alert

+++ src/app/admin/post/index/index.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { PostService } from '../post.service';
 import { Post } from '../post';
 
+import Swal from 'sweetalert2/dist/sweetalert2.js'
+
 @Component({
   selector: 'app-index',
   templateUrl: './index.component.html',
@@ -20,10 +22,37 @@ export class IndexComponent implements OnInit {
     })
   }
 
+  simpleAlert() {
+    Swal.fire('Hello world!');
+  }
+
   deletePost(id) {
-    this.postService.delete(id).subscribe(res => {
-      this.posts = this.posts.filter(item => item.id !== id);
-      console.log('Post deleted successfully!');
+    Swal.fire({
+      title: 'Are you sure want to remove?',
+      text: 'You will not be able to recover this file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        this.postService.delete(id).subscribe(res => {
+          this.posts = this.posts.filter(item => item.id !== id);
+          console.log('Post deleted successfully!');
+          Swal.fire(
+            'Deleted!',
+            'Your imaginary file has been deleted.',
+            'success'
+          )
+        })
+
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
     })
   }
